+++ app/src/main/java/com/joshuarichardson/fivewaystowellbeing/storage/dao/DatabaseInsertionHelper.java
@@ -8,31 +8,31 @@ import java.util.ArrayList;
 
 public class DatabaseInsertionHelper {
 
-    public static ArrayList<Integer> insert(SurveyResponseElement[] surveyResponseElements, SurveyResponseElementDao surveyResponseElementDao) {
-        ArrayList<Integer> surveyResponseElementIdList = new ArrayList<>();
+    public static ArrayList<Long> insert(SurveyResponseElement[] surveyResponseElements, SurveyResponseElementDao surveyResponseElementDao) {
+        ArrayList<Long> surveyResponseElementIdList = new ArrayList<>();
 
         for (SurveyResponseElement surveyResponseElement : surveyResponseElements) {
-            surveyResponseElementIdList.add((int) surveyResponseElementDao.insert(surveyResponseElement));
+            surveyResponseElementIdList.add(surveyResponseElementDao.insert(surveyResponseElement));
         }
 
         return surveyResponseElementIdList;
     }
 
-    public static ArrayList<Integer> insert(ActivityRecord[] activityRecords, ActivityRecordDao activityRecordDao) {
-        ArrayList<Integer> activityRecordIdList = new ArrayList<>();
+    public static ArrayList<Long> insert(ActivityRecord[] activityRecords, ActivityRecordDao activityRecordDao) {
+        ArrayList<Long> activityRecordIdList = new ArrayList<>();
 
         for (ActivityRecord activityRecord : activityRecords) {
-            activityRecordIdList.add((int) activityRecordDao.insert(activityRecord));
+            activityRecordIdList.add(activityRecordDao.insert(activityRecord));
         }
 
         return activityRecordIdList;
     }
 
-    public static ArrayList<Integer> insert(SurveyResponse[] surveyResponses, SurveyResponseDao surveyResponseDao) {
-        ArrayList<Integer> surveyResponseIdList = new ArrayList<>();
+    public static ArrayList<Long> insert(SurveyResponse[] surveyResponses, SurveyResponseDao surveyResponseDao) {
+        ArrayList<Long> surveyResponseIdList = new ArrayList<>();
 
         for (SurveyResponse surveyResponse : surveyResponses) {
-            surveyResponseIdList.add((int) surveyResponseDao.insert(surveyResponse));
+            surveyResponseIdList.add(surveyResponseDao.insert(surveyResponse));
         }
 
         return surveyResponseIdList;
